Move plotter -> plotters/base
